#  lib/db: Optimize memory usage for huge files, refactor updating and fix benchs