This started in a discussion about memory use spiking when scanning is finished here: https://forum.syncthing.net/t/out-of-memory-crash-while-scanning/11170/13
The I thought I saw unnecessary block copies abound, but it turned out I was wrong. However in the process I saw other potential memory optimizations and some refactor potential:

 - The function `(m *BlockMap) Add` is only used in test and there it can be replaced by `Update` -> remove
 - Do not store existing files to be discarded in a list -> move discarding into `(m *BlockMap) Update`.  
This might also be achievable while keeping `Discard` separate, but I quite like having a single `Update` function as they are never used alone and it's the equivalent function to `Update` on `FileSet`.
 - Check whether the block map batch exceeds the maximum size after every block instead of after every file (i.e. for one large file, the transaction will be split into several batches).
 - Discard block data early when removing an old file from the block map.

### Testing

I fixed the previously broken benchmarks (unnoticed due to build tags). There is no consistent change visible between master and this PR - so this may save some memory and should not impact performance (much).