@calmh 2018-01-20
How about a memory benchmark? (Using `go test -bench . -benchmem` or the t.ReportAllocs() or whatever the method is called, I forget.)