@imsodin 2018-01-20
I increase benchtime to get better results. Allocs/op seem to correlate with benchtime and vary slightly by run (shouldn't they be deterministic?). If there is a trend, then weirdly enough it looks like overall the PR uses slightly more allocs/op.
 
before:
```
BenchmarkReplaceAll-4                         10	2572511544 ns/op	253095833 B/op	  606398 allocs/op
BenchmarkUpdateOneChanged-4     	  300000	    120480 ns/op	   87763 B/op	     548 allocs/op
BenchmarkUpdateOneUnchanged-4   	  200000	    119976 ns/op	   87698 B/op	     549 allocs/op
BenchmarkUpdateOneFile-4        	  200000	    215976 ns/op	   34519 B/op	     232 allocs/op
```

after:
```
BenchmarkReplaceAll-4           	      20	2405571755 ns/op	268937934 B/op	  636290 allocs/op
BenchmarkUpdateOneChanged-4     	  200000	    142700 ns/op	  108597 B/op	     558 allocs/op
BenchmarkUpdateOneUnchanged-4   	  200000	    131637 ns/op	  108556 B/op	     559 allocs/op
BenchmarkUpdateOneFile-4        	  200000	    232549 ns/op	   36151 B/op	     267 allocs/op
```

I must admit I didn't look at what the benchmarks do besides what was needed to get the to run. Probably there is none for the single big file scenario.