@calmh 2018-01-27
So not to be annoying, but the benchmark changes (to the extent they're statistically relevant) all became worse?