@imsodin 2018-01-27
Not annoying at all, I kind of purposely forgot about this PR :P  
I wouldn't call the bench changes statistically relevant at all - most of them are clearly not relevant and in those which may be, I observed weird dependencies on runtime parameters. I didn't see enough potential gain to go through the work of either writing "better" benchmarks (whatever that would be) or do a actual statistics on benchmark results.

So I close this and will do a new PR with the benchmarks fixes and maybe at some point I'll come back on some of the changes here.